backend/requirements.txt
fastapi==0.128.0
pydantic==2.12.5
SQLAlchemy==2.0.46
gunicorn
psycopg2-binary
uvicorn

passlib[bcrypt]
python-jose[cryptography]
python-multipart
openpyxl
pandas
python-dotenv